+++ player/player.py
@@ -2,7 +2,7 @@ from items.property import Property
 from player import player_card as pc
 import universal.game_board as gb
 import fields.fcontainer as fc
-import rules.rule_ui as r_ui
+import universal.side_bar as sb
 import pygame
 import math
 
@@ -43,7 +43,7 @@ class Player:
 
             # Clear the screen (or redraw the background)
             screen.fill("purple")
-            r_ui.rule_ui_setup(screen)
+            sb.sb_setup(screen)
             pc.player_card(screen,self)
             pc.win_condition_Card(screen, self)
             for fld in fc.f_container: