+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This build is not intended for submission, but rather for the developers to keep
 
 - Some code has been restructured.
 - That means a fix for the player logic had to be made.
+- The rule UI file is now the sidebar file in `universal`.
 
 ## Known issues